Initial commit of all the classes.
These classes invoice:
Invoice.java - Class for the invoices, to be populated by the database.
InvoiceViewController.java - Class to control invoice-view.fxml
Item.java - Class for the items, to be populated by the database.
Status.java - Enum for the invoice Statuses.
invoice_view.fxml - FXML page for the invoice views.